tests/minio/requirements.txt
minio
python-dotenv